fix namespace and tests
