+++ tests/Stubs/StubService.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Tests\Stubs;
+
+final class StubService implements StubServiceInterface
+{
+}
